5000: divide trigger form in multi components
